add casenotes standardisation logic to gateway
